art_and_logic/2-epen/requirements.txt
tox==3.14.3
virtualenv==20.0.1
python-dotenv==0.11.0
numpy==1.18.1
